backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
